backend/requirements.txt
fastapi[all]
pydantic
pydantic-settings
python-dotenv
sqlalchemy
alembic
psycopg2-binary